kiosk: chromium exec options in cage
Chromium needed to be able to write to the CWD. Similar issue was
reported:
https://github.com/NixOS/nixpkgs/issues/229235#issuecomment-1891060190

reg url is an iframe and needs to have some of the more secure options
in chromium disabled.

Disabling the GPU for good measure since we were seeing errors in the
logs related to rpi GPU and wayland.